Move the schema validation test to the new client-server test module
