+++ integration-tests/client-server/pom.xml
@@ -48,6 +48,7 @@
                 <executions>
                     <execution>
                         <goals>
+                            <goal>generate-code</goal>
                             <goal>build</goal>
                         </goals>
                     </execution>
