+++ integration-tests/client-server/src/main/java/io/quarkiverse/cxf/it/server/xml/schema/validation/CalculatorServiceImpl.java
