+++ integration-tests/client-server/src/main/resources/application.properties
@@ -5,3 +5,13 @@ quarkus.cxf.endpoint."/addressing-anonymous".implementor = io.quarkiverse.cxf.it
 quarkus.cxf.endpoint."/addressing-decoupled".implementor = io.quarkiverse.cxf.it.ws.addressing.server.decoupled.WsAddressingImpl
 
 quarkus.native.resources.includes = wsdl/*.wsdl
+
+# XMl Schema validation
+quarkus.cxf.codegen.wsdl2java.includes = wsdl/*.wsdl
+quarkus.cxf.codegen.wsdl2java.package-names = io.quarkiverse.cxf.it.server.xml.schema.validation.model
+quarkus.cxf.codegen.wsdl2java.wsdl-location = classpath:wsdl/calculator.wsdl
+quarkus.cxf.endpoint."/schema-validated-calculator".implementor = io.quarkiverse.cxf.it.server.xml.schema.validation.CalculatorServiceImpl
+# https://github.com/quarkiverse/quarkus-cxf/issues/557 @WebServide(wsdlLocation = "wsdl/calculator.wsdl") does not work
+quarkus.cxf.endpoint."/schema-validated-calculator".wsdl = wsdl/calculator.wsdl
+quarkus.cxf.endpoint."/schema-validated-calculator".logging.enabled = true
+quarkus.native.resources.includes = wsdl/*
