+++ integration-tests/client-server/src/main/resources/wsdl/calculator.xsd
