+++ integration-tests/server/src/main/resources/application.properties
@@ -41,16 +41,6 @@ quarkus.cxf.endpoint."/soap-message".implementor = io.quarkiverse.cxf.it.server.
 quarkus.cxf.endpoint."/source-message".implementor = io.quarkiverse.cxf.it.server.provider.SourceMessageProvider
 quarkus.cxf.endpoint."/stream-source-payload".implementor = io.quarkiverse.cxf.it.server.provider.StreamSourcePayloadProvider
 
-
-# XMl Schema validation
-quarkus.cxf.codegen.wsdl2java.includes = wsdl/*.wsdl
-quarkus.cxf.codegen.wsdl2java.package-names = io.quarkiverse.cxf.it.server.xml.schema.validation.model
-quarkus.cxf.codegen.wsdl2java.wsdl-location = classpath:wsdl/calculator.wsdl
-quarkus.cxf.endpoint."/schema-validated-calculator".implementor = io.quarkiverse.cxf.it.server.xml.schema.validation.CalculatorServiceImpl
-# https://github.com/quarkiverse/quarkus-cxf/issues/557 @WebServide(wsdlLocation = "wsdl/calculator.wsdl") does not work
-quarkus.cxf.endpoint."/schema-validated-calculator".wsdl = wsdl/calculator.wsdl
-quarkus.cxf.endpoint."/schema-validated-calculator".logging.enabled = true
-quarkus.native.resources.includes = wsdl/*
 #
 # CLIENTS
 #
